requirements.txt
tqdm==4.45.0
requests==2.23.0
praw==6.5.1
Pillow==7.1.2
ffmpeg-python==0.2.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-api-python-client==1.8.2
#pytesseract==0.3.4
